feat(jnml): adds -lems-graph flag to jnml
This is primarily added for clarity, to stress that this produces a
summary graph for the simulation from its LEMS description, rather than
one of the model description from the nml file (which is not implemented
here but resides in pynml).

The usage message has also been updated to note that `-graph` is an
alias of `-lems-graph`.

Fixes https://github.com/NeuroML/jNeuroML/issues/73